requirements.txt
flask==3.0.3
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
tavily-python==0.3.3